backendMain/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
beautifulsoup4==4.14.2
billiard==4.2.3
bs4==0.0.2
celery==5.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==46.0.3
Django==5.1.14
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.9.0
django-ninja==1.5.0
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
greenlet==3.2.4
gunicorn==23.0.0
hiredis==2.3.2
idna==3.11
kombu==5.6.0
Levenshtein==0.27.3
lxml==6.0.2
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
playwright==1.56.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
PyJWT==2.10.1
pypdfium2==5.1.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
python-multipart==0.0.20
RapidFuzz==3.14.3
redis==5.2.0
requests==2.32.5
schedule==1.2.2
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
whitenoise==6.11.0
python-json-logger==2.0.7
